Move fixtures to conftest so tests don't have to manually import them
