+++ tests/unit/grpc_internals/zeebe_message_adapter_test.py
@@ -2,11 +2,13 @@ from random import randint
 from unittest.mock import MagicMock
 from uuid import uuid4
 
+import grpc
+import pytest
 from zeebe_grpc.gateway_pb2 import *
 
 from pyzeebe.exceptions import MessageAlreadyExists
 from pyzeebe.grpc_internals.zeebe_message_adapter import ZeebeMessageAdapter
-from tests.unit.utils.grpc_utils import *
+from tests.unit.utils.grpc_utils import GRPCStatusCode
 from tests.unit.utils.random_utils import RANDOM_RANGE
 
 zeebe_message_adapter: ZeebeMessageAdapter
